requirements.txt
# Django and DRF
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.12.0
django-filter==23.5

# PostgreSQL
psycopg2-binary==2.9.9

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# API Documentation
drf-spectacular==0.27.0

# Database
sqlalchemy==2.0.23

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Testing and Development
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
flake8==6.1.0
black==23.12.1
isort==5.13.2
pylint==3.0.3

# Security
django-ratelimit==4.1.0

# Protocol Buffers
protobuf==4.25.8
grpcio-tools==1.60.0
